requirements.txt
# Core dependencies
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
requests>=2.28.0
python-whois>=0.9.0
lightgbm>=4.0.0
xgboost>=1.7.0

# Optional dependencies for enhanced filtering
django-filter>=23.0.0

# Database (optional - for production)
psycopg2-binary>=2.9.0



